+++ FABLECRAFT NUXT BUILD/components/TestComponent.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="bg-red-500 text-white p-4 m-4">
+    <h1>TEST COMPONENT WORKS!</h1>
+    <p>This is a test to see if components are loading.</p>
+  </div>
+</template>
+
+<script setup lang="ts">
+console.log('TEST COMPONENT LOADED')
+</script>
